The m/z data type has been shifted from float to double. This allows to handle high precision mass spectrometric data too.

+++ openchrom/plugins/net.openchrom.chromatogram.msd.converter.supplier.cdf/src/net/openchrom/chromatogram/msd/converter/supplier/cdf/io/support/IScanSupport.java
@@ -39,15 +39,15 @@ public interface IScanSupport {
 	 * Returns the min ion of the given scan.
 	 * 
 	 * @param scan
-	 * @return float
+	 * @return double
 	 */
-	float getMinIon(int scan);
+	double getMinIon(int scan);
 
 	/**
 	 * Returns the max ion of the given scan.
 	 * 
 	 * @param scan
-	 * @return float
+	 * @return double
 	 */
-	float getMaxIon(int scan);
+	double getMaxIon(int scan);
 }
